+++ R/SDistribution_Arcsine.R
@@ -134,19 +134,9 @@ Arcsine <- R6Class("Arcsine",
     #' @description
     #' Sets the value(s) of the given parameter(s).
     setParameterValue = function(..., lst = NULL, error = "warn") {
-      if (is.null(lst)) {
-        lst <- list(...)
-      }
-      if ("lower" %in% names(lst) & "upper" %in% names(lst)) {
-        checkmate::assert(lst[["lower"]] <= lst[["upper"]])
-      } else if ("lower" %in% names(lst)) {
-        checkmate::assert(lst[["lower"]] <= self$getParameterValue("upper"))
-      } else if ("upper" %in% names(lst)) {
-        checkmate::assert(lst[["upper"]] >= self$getParameterValue("lower"))
-      }
-
-      super$setParameterValue(lst = lst, error = error)
-      private$.properties$support <- Interval$new(self$getParameterValue("lower"), self$getParameterValue("upper"))
+      super$setParameterValue(..., lst = lst, error = error)
+      private$.properties$support <- Interval$new(self$getParameterValue("lower"),
+                                                  self$getParameterValue("upper"))
       invisible(self)
     }
   ),